ai-service/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# AI/ML Core
torch==2.5.1
transformers==4.36.2
langchain==0.1.0
langchain-community==0.0.10
llama-cpp-python==0.2.27

# Hugging Face (not needed for Groq, but keeping for flexibility)
huggingface-hub==0.20.2
accelerate==0.25.0
bitsandbytes==0.41.3

# Groq API for Llama inference
groq==0.37.1

# NLP & Text Processing
spacy==3.7.2
nltk==3.8.1
sentencepiece==0.1.99

# Azure Services
azure-ai-formrecognizer>=3.3.0
azure-ai-translation-text>=1.0.0
azure-identity>=1.15.0

# Database & Caching
redis==5.0.1
pymongo==4.6.1

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0  # For testing API endpoints

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# OCR & Document Processing (Optional for Phase 2)
# pytesseract==0.3.10
# pdf2image==1.17.0
# Pillow==10.2.0
